# admin-service/docs/swagger.yaml
definitions:
  models.AppRequest:
    properties:
      name:
        type: string
    type: object
  models.AppResponse:
    properties:
      secret:
        type: string
    type: object
  models.Error:
    properties:
      code:
        example: 0
        type: integer
      message:
        example: status
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.LoginResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  models.RegisterResponse:
    properties:
      message:
        type: string
    type: object
  models.Rule:
    properties:
      match_type:
        type: string
      parameter:
        type: string
    type: object
  models.RuleRequest:
    properties:
      app_id:
        type: integer
      rule:
        $ref: '#/definitions/models.Rule'
    type: object
info:
  contact:
    email: hacksparrow169@gmail.com
    name: topgs
  description: Admin Panel for trinit Application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: trinit Admin
  version: "1.0"
paths:
  /v1/app/deleterule/{id}:
    delete:
      consumes:
      - application/json
      description: Delete rules
      parameters:
      - description: rule id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete rules
      tags:
      - App
  /v1/app/getrules/{id}:
    get:
      consumes:
      - application/json
      description: Get rules
      parameters:
      - description: app id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get rules
      tags:
      - App
  /v1/app/registerapp:
    post:
      consumes:
      - application/json
      description: Register app
      parameters:
      - description: Regsiter app
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/models.AppRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AppResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Register app
      tags:
      - App
  /v1/app/registerrule:
    post:
      consumes:
      - application/json
      description: Register rule
      parameters:
      - description: Regsiter app
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/models.RuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AppResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Register rule
      tags:
      - App
  /v1/app/updaterule/{id}:
    put:
      consumes:
      - application/json
      description: Update rule
      parameters:
      - description: updated rule
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/models.RuleRequest'
      - description: rule id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Update rule
      tags:
      - App
  /v1/user/signin:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: Authenticate user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Login user
      tags:
      - User
  /v1/user/signup:
    post:
      consumes:
      - application/json
      description: register an user
      parameters:
      - description: Add user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Register an user
      tags:
      - User
  /v1/user/verifyemail/{verificationCode}:
    post:
      consumes:
      - application/json
      description: verify an user by sending email
      parameters:
      - description: Verify user
        in: path
        name: verificationCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: verify an user
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"